No More Ghost Cells
I have written the routine to create the Backward-Euler operator. It no longer assumes that the concentrations include ghost cells. The flux at the left boundary is now written as

K*C_bnd

where K is a parameter passed to the routine and C_bnd is approximated with linear extrapolation from the first two interior points.

The left boundary condition also needs to be added to the rhs vector in a particular way. Testing for zero flux boundary conditions has been done. Now I need to make sure it works for K != 0.